+++ src/views/From.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useState } from 'react';
+import React from 'react';
 import { useDispatch } from 'react-redux';
 import Radio from '@material-ui/core/Radio';
 import RadioGroup from '@material-ui/core/RadioGroup';
@@ -10,69 +10,42 @@ import { PrimaryButton, TextInput } from '../components/UI';
 import '../assets/css/common.css';
 import { register } from '../reducks/troubleLists/operations';
 import { FORM_ERROR, FORM_SUCCESS } from '../common/messages/ui/FormResults';
-import { slackWebhook } from '../reducks/commons/operatirons';
+// import { slackWebhook } from '../reducks/commons/operatirons';
+import { useForms } from '../hooks/useForms';
 
 const From: React.FC = () => {
   const dispatch = useDispatch();
-  const [username, setUsername] = useState('');
-  const [age, setAge] = useState('0');
-  const [gender, setGender] = useState('男');
-  const [trouble, setTrouble] = useState('');
-  const [backGround, setBackGround] = useState('');
-  const [remark, setRemark] = useState('');
-
-  const [usernameError, setUsernameError] = useState('');
-  const [troubleError, setTroubleError] = useState('');
-  const [backGroundError, setBackGroundError] = useState('');
-
-  const inputUsername = useCallback(
-    (event: React.ChangeEvent<HTMLInputElement>) => {
-      setUsername(event.target.value);
-    },
-    [setUsername]
-  );
-
-  const inputAge = useCallback(
-    (event: React.ChangeEvent<HTMLInputElement>) => {
-      setAge(event.target.value);
-    },
-    [setAge]
-  );
-
-  const inputGender = useCallback(
-    (event: React.ChangeEvent<HTMLInputElement>) => {
-      setGender(event.target.value);
-    },
-    [setGender]
-  );
-
-  const inputBackGround = useCallback(
-    (event: React.ChangeEvent<HTMLInputElement>) => {
-      setBackGround(event.target.value);
-    },
-    [setBackGround]
-  );
-
-  const inputTrouble = useCallback(
-    (event: React.ChangeEvent<HTMLInputElement>) => {
-      setTrouble(event.target.value);
-    },
-    [setTrouble]
-  );
-
-  const inputRemark = useCallback(
-    (event: React.ChangeEvent<HTMLInputElement>) => {
-      setRemark(event.target.value);
-    },
-    [setRemark]
-  );
+  const {
+    username,
+    age,
+    gender,
+    trouble,
+    backGround,
+    remark,
+    usernameError,
+    troubleError,
+    backGroundError,
+    inputUsername,
+    inputAge,
+    inputGender,
+    inputTrouble,
+    inputBackGround,
+    inputRemark,
+    setUsernameError,
+    setTroubleError,
+    setBackGroundError,
+    resetForm,
+    getErrorMesssage,
+  } = useForms();
 
   const submit = async (): Promise<void> => {
-    const userErrorMsg = !username ? '必須項目です' : '';
+    const {
+      userErrorMsg,
+      troubleErrorMsg,
+      backGroundErrorMsg,
+    } = getErrorMesssage(username, trouble, backGround);
     setUsernameError(userErrorMsg);
-    const troubleErrorMsg = !trouble ? '必須項目です' : '';
     setTroubleError(troubleErrorMsg);
-    const backGroundErrorMsg = !backGround ? '必須項目です' : '';
     setBackGroundError(backGroundErrorMsg);
 
     if (userErrorMsg || troubleErrorMsg || backGroundErrorMsg) {
@@ -88,34 +61,30 @@ const From: React.FC = () => {
 
       window.scrollTo(0, 0);
 
-      setUsername('');
-      setAge('0');
-      setTrouble('');
-      setBackGround('');
-      setRemark('');
+      resetForm();
     } catch (e) {
       alert(FORM_ERROR);
       return;
     }
 
-    const url =
-      'https://hooks.slack.com/services/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';
-    const payload = {
-      text: `【投稿者名】
-      ${username}
-      【年齢】
-      ${age}
-      【性別】
-      ${gender}
-      【悩み】
-      ${trouble}
-      【背景】
-      ${backGround}
-      【備考】
-      ${remark}`,
-    };
+    // const url =
+    //   'https://hooks.slack.com/services/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';
+    // const payload = {
+    //   text: `【投稿者名】
+    //   ${username}
+    //   【年齢】
+    //   ${age}
+    //   【性別】
+    //   ${gender}
+    //   【悩み】
+    //   ${trouble}
+    //   【背景】
+    //   ${backGround}
+    //   【備考】
+    //   ${remark}`,
+    // };
 
-    await dispatch(slackWebhook(url, payload));
+    // await dispatch(slackWebhook(url, payload));
     alert(FORM_SUCCESS);
     dispatch(push('/'));
   };
